Latest change from main to Rico

+++ back-office/server/src/main/java/proj/cloud/ath/controllers/api/v1/CarModelController.java
@@ -6,12 +6,14 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import proj.cloud.ath.entities.CarModel;
 import proj.cloud.ath.response.RestApiResponse;
 import proj.cloud.ath.services.CarModelService;
+import proj.cloud.ath.utils.JwtUtil;
 
 @RestController
 @RequestMapping("/api/v1/car-models")
@@ -20,31 +22,62 @@ public class CarModelController {
     @Autowired
     private CarModelService service;
 
+    @Autowired
+    private JwtUtil jwtUtil;
+
     @GetMapping
-    public RestApiResponse findAll() {
-        return new RestApiResponse(service.findAll(), 200);
+    public RestApiResponse findAll(@RequestHeader(value = "Authorization") String bearerToken ) {
+        String token = bearerToken.substring(7);
+        RestApiResponse response = new RestApiResponse();
+        if (jwtUtil.isValidToken(token)) {
+            response = new RestApiResponse(service.findAll(), 200);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
+        return response;
     }
 
     @GetMapping("/{id}")
-    public RestApiResponse findById(@PathVariable Long id) {
-        return new RestApiResponse(service.findById(id), 200);
+    public RestApiResponse findById(@RequestHeader(value = "Authorization") String bearerToken, @PathVariable Long id) {
+        String token = bearerToken.substring(7);
+        RestApiResponse response = new RestApiResponse();
+        if (jwtUtil.isValidToken(token)) {
+            response = new RestApiResponse(service.findById(id), 200);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
+        return response;
     }
 
     @PostMapping
-    public RestApiResponse create(@RequestBody CarModel carModel) {
+    public RestApiResponse create(@RequestHeader(value = "Authorization") String bearerToken,@RequestBody CarModel carModel) {
+        String token = bearerToken.substring(7);
         RestApiResponse response = new RestApiResponse();
-        carModel.setState(0);
-        service.save(carModel);
-        response.setPayload(carModel);
-        response.setStatus(201);
+        if (jwtUtil.isValidToken(token)) {
+            carModel.setState(0);
+            service.save(carModel);
+            response.setPayload(carModel);
+            response.setStatus(201);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
         return response;
     }
 
     @DeleteMapping("/{id}")
-    public RestApiResponse delete(@PathVariable Long id) {
+    public RestApiResponse delete(@RequestHeader(value = "Authorization") String bearerToken, @PathVariable Long id) {
+        String token = bearerToken.substring(7);
         RestApiResponse response = new RestApiResponse();
-        service.delete(id);
-        response.setStatus(200);
+        if (jwtUtil.isValidToken(token)) {
+            service.delete(id);
+            response.setStatus(200);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
         return response;
     }
 }
