+++ back-office/server/src/main/java/proj/cloud/ath/controllers/api/v1/TransmissionTypeController.java
@@ -7,10 +7,12 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import proj.cloud.ath.services.TransmissionTypeService;
+import proj.cloud.ath.utils.JwtUtil;
 import proj.cloud.ath.entities.TransmissionType;
 import proj.cloud.ath.response.RestApiResponse;
 
@@ -20,45 +22,83 @@ public class TransmissionTypeController {
 
     @Autowired
     private TransmissionTypeService service;
+    
+    @Autowired
+    private JwtUtil jwtUtil;
 
     @GetMapping
-    public RestApiResponse findAll() {
-        return new RestApiResponse(service.findAll(), 200);
+    public RestApiResponse findAll(@RequestHeader(value = "Authorization") String bearerToken ) {
+        String token = bearerToken.substring(7);
+        RestApiResponse response = new RestApiResponse();
+        if (jwtUtil.isValidToken(token)) {
+            response = new RestApiResponse(service.findAll(), 200);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
+        return response;
     }
 
     @GetMapping("/{id}")
-    public RestApiResponse findById(@PathVariable Long id) {
-        return new RestApiResponse(service.findById(id), 200);
+    public RestApiResponse findById(@RequestHeader(value = "Authorization") String bearerToken, @PathVariable Long id) {
+        String token = bearerToken.substring(7);
+        RestApiResponse response = new RestApiResponse();
+        if (jwtUtil.isValidToken(token)) {
+            response = new RestApiResponse(service.findById(id), 200);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
+        return response;
     }
 
     @PostMapping
-    public RestApiResponse create(@RequestBody TransmissionType transmissionType) {
+    public RestApiResponse create(
+            @RequestHeader(value = "Authorization") String bearerToken, @RequestBody TransmissionType transmissionType) {
+        String token = bearerToken.substring(7);
         RestApiResponse response = new RestApiResponse();
-        service.save(transmissionType);
-        response.setPayload(transmissionType);
-        response.setStatus(201);
+        if (jwtUtil.isValidToken(token)) {
+            service.save(transmissionType);
+            response.setPayload(transmissionType);
+            response.setStatus(201);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
         return response;
     }
 
     @PutMapping
-    public RestApiResponse update(@RequestBody TransmissionType transmissionType) {
+    public RestApiResponse update(@RequestHeader(value = "Authorization") String bearerToken,@RequestBody TransmissionType transmissionType) {
+        String token = bearerToken.substring(7);
         RestApiResponse response = new RestApiResponse();
-        response.setStatus(400);
-        TransmissionType oldTransmissionType = service.findById(transmissionType.getId());
-        if (oldTransmissionType != null) {
-            oldTransmissionType.setName(transmissionType.getName());
-            service.save(oldTransmissionType);
-            response.setStatus(200);
-            response.setPayload(oldTransmissionType);
+        if (jwtUtil.isValidToken(token)) {
+            response.setStatus(400);
+            TransmissionType oldTransmissionType = service.findById(transmissionType.getId());
+            if (oldTransmissionType != null) {
+                oldTransmissionType.setName(transmissionType.getName());
+                service.save(oldTransmissionType);
+                response.setStatus(200);
+                response.setPayload(oldTransmissionType);
+            }
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
         }
         return response;
     }
 
     @DeleteMapping("/{id}")
-    public RestApiResponse delete(@PathVariable Long id) {
+    public RestApiResponse delete(@RequestHeader(value = "Authorization") String bearerToken, @PathVariable Long id) {
+        String token = bearerToken.substring(7);
         RestApiResponse response = new RestApiResponse();
-        service.delete(id);
-        response.setStatus(200);
+        if (jwtUtil.isValidToken(token)) {
+            service.delete(id);
+            response.setStatus(200);
+        } else {
+            response.setMessage("Invalid token");
+            response.setStatus(403);
+        }
         return response;
     }
 }
